backend/requirements.txt
# Flask 相关
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.10.1
Flask-Mail==0.10.0
python-dotenv==1.0.1
alibabacloud-oss-v2
pillow
pymysql